feat: add new parser

+++ src/leaf/main.py
@@ -1,6 +1,7 @@
 import argparse
 import shutil
 from pathlib import Path
+from time import perf_counter
 
 from loguru import logger
 
@@ -13,6 +14,8 @@ def main():
     """Main EOP"""
 
     args, other_args = parse_args()
+    logger.trace(args)
+    logger.trace(other_args)
 
     if args.task == "generate":
         run_generator(args)
@@ -33,11 +36,12 @@ def parse_args():
     subparsers = cli.add_subparsers(dest="task", required=True)
 
     generate_parser = subparsers.add_parser("generate")
-    generate_parser.add_argument("--manifests-file", required=True)
+    generate_parser.add_argument("--manifests-file", default="manifests.txt")
     # generate_parser.add_argument("--output-path")
+    generate_parser.add_argument("--overwrite", action="store_true")
 
     download_parser = subparsers.add_parser("download")
-    download_parser.add_argument("--manifests-file", required=True)
+    download_parser.add_argument("--manifests-file", default="manifests.txt")
     download_parser.add_argument("--output-path")
 
     return cli.parse_known_args()
@@ -46,38 +50,44 @@ def parse_args():
 def run_generator(args: argparse.Namespace):
     manifests_file = Path(args.manifests_file)
     # output_path = util.get_path(args.output_path, default=OUT_PATH)
+    overwrite = bool(args.overwrite)
 
     if not OUT_PATH.exists():
         logger.debug("Creating output path because it doesn't exist")
         OUT_PATH.mkdir(parents=True, exist_ok=True)
 
-    if MANIFESTS_PATH.exists():
+    if overwrite and MANIFESTS_PATH.exists():
         logger.warning("Found previously-created files in generated. Removing all of them")
         shutil.rmtree(MANIFESTS_PATH)
 
-    if INDEXES_PATH.exists():
+    if overwrite and INDEXES_PATH.exists():
         logger.warning("Found previously-created files in indexes. Removing all of them")
         shutil.rmtree(INDEXES_PATH)
 
     MANIFESTS_PATH.mkdir(parents=True, exist_ok=True)
     INDEXES_PATH.mkdir(parents=True, exist_ok=True)
 
+    start = perf_counter()
     with open(manifests_file, "r", encoding="utf-8") as f:
         for line in f:
+            line = line.strip()
+
             # Skip commented lines for testing
-            if line.startswith("//"):
+            if len(line) == 0 or line.startswith("//"):
                 continue
 
             data = line.split(":")
             steam_info = SteamInfo(
-                data[0].strip(),
-                data[1].strip(),
-                data[2].strip(),
-                (data[3] if len(data) > 3 else "public").strip(),
+                data[0],
+                data[1],
+                data[2],
+                (data[3] if len(data) > 3 else "public"),
             )
 
             logger.trace("Parsed manifest entry with info:\n\t{}", steam_info.to_json())
-            generator.generate_manifests(steam_info)
+            generator.generate(steam_info, overwrite=overwrite)
+    stop = perf_counter()
+    logger.info(f"All manifests generated! Time taken: {((stop - start) * 1000):.3f}ms")
 
 
 def run_downloader(args: argparse.Namespace, other_args: list[str]):
